Replace rotation meter with altimeter
Measuring exact rotation speed isn't very useful and a bit misleading
since the camera rotates.

The way I'm currently measuring altitude is pretty inefficient though...

Also, I already have existing code that computes something similar to
altitude.  To determine zoom, I calculate "separation", which is a rough
approximation of the minimum distance to the asteroid.  This is not the
same as "altitude" in the sense of the new method that computes minimum
distance to the ground in the direction of the center of mass.  For now
I think I'll leave it around, because I don't think I even want to use
"altitude" in place of "separation", and "altitude" is more inefficient
to calculate.  However, if I cache the result of getAltitude, it might
be a good idea to try replacing "separation" with "altitude".  But doing
caching sounds harder for getAltitude than for getTransform or getBounds
because technically the asteroid could move too, not just the ship.